[Northamptonshire] Move assets to configuration.
