+++ templates/web/fixmystreet.com/footer_extra_js.html
@@ -5,7 +5,6 @@ scripts.push(
 IF bodyclass.match('mappage') AND NOT bodyclass.match('accountpage');
     scripts.push( version('/cobrands/fixmystreet/assets.js') );
     scripts.push( version('/cobrands/buckinghamshire/grass_cutting.js') );
-    scripts.push( version('/cobrands/northamptonshire/assets.js') );
     scripts.push( version('/cobrands/oxfordshire/assets.js') );
     scripts.push( version('/cobrands/westminster/assets.js') );
     scripts.push( version('/cobrands/tfl/assets.js') );
