+++ templates/web/northamptonshire/footer_extra_js.html
@@ -1,1 +1,1 @@
-[% PROCESS 'footer_extra_js_base.html' highways=1 cobrand_js=1 validation=1 roadworks=1 %]
+[% PROCESS 'footer_extra_js_base.html' highways=1 validation=1 roadworks=1 %]
